requirements.txt
django
django-rest-framework
djangorestframework-simplejwt
django-filter